pyproject.toml
[tool.poetry]
name = "podaac-agents"
version = "0.1.0"
description = "A growing collection of intelligent AI agents designed to assist with Earth science data discovery, access, and analysis."
authors = ["jamesfwood"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/podaac/podaac-agents"
packages = [{include = "podaac_agents", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
strands-agents = "^1.14.0"
strands-agents-tools = "^0.2.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
